HAI-2310 Moved application entity and data to hakemus
